Move OpeningPeriod to test code
